refactor: argument resolver 위치 변경
